requirements.txt
aioredis==1.3.1
amqp==5.1.1
gunicorn==20.1.0
uvicorn==0.20.0
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.1
Automat==22.10.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
channels==3.0.3
channels-redis==3.2.0
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
daphne==3.0.2
Django==4.1.5
django-cors-headers==3.13.0
django-environ==0.9.0
django-split-settings==1.2.0
djangorestframework==3.14.0
docopt==0.6.2
drf-jwt==1.19.2
drf-yasg==1.21.4
flake8==6.0.0
hiredis==2.1.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
mccabe==0.7.0
msgpack==1.0.4
packaging==23.0
pipreqs==0.4.11
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.6.0
pyOpenSSL==23.0.0
pytz==2022.7
redis==4.4.2
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.10.0
txaio==23.1.1
typing-extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.5
yarg==0.1.9
zope.interface==5.5.2
daphne==3.0.2